Add legacy and fpx api manager
Fix issue with duplicate recent workspace entries
