+++ fpx-app/Cargo.toml
@@ -20,21 +20,29 @@ axum = { workspace = true, default-features = false, features = [
     "tokio",
     "tracing",
     "ws",
-] }
+], optional = true }
 fpx = { version = "0.1.0", path = "../fpx", features = ["config", "libsql"] }
+nix = { version = "0.29", default-features = false, features = ["signal"] }
 schemars = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tauri = { version = "2.0.0-rc.14", features = ["config-toml"] }
 tauri-plugin-dialog = "2.0.0-rc"
 tauri-plugin-store = { version = "2.0.0-rc" }
-tokio = { version = "1", default-features = false, features = ["sync"] }
+tokio = { version = "1", default-features = false, features = [
+    "sync",
+    "process",
+    "macros",
+] }
 tracing = { version = "0.1" }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 
 [features]
 # This feature is used for production builds or when a dev server is not specified, DO NOT REMOVE!!
 custom-protocol = ["tauri/custom-protocol"]
+
+# This feature enables the new Rust API. If not enabled it will use the TS API.
+fpx-api = ["dep:axum"]
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
 tauri-plugin-window-state = "2.0.0-rc"
