+++ fpx-app/Tauri.toml
@@ -3,7 +3,7 @@ identifier = "com.fiberplane.fpx"
 [build]
 devUrl = "http://localhost:5173"
 beforeBuildCommand = "pnpm build:fpx-studio"
-beforeDevCommand = "pnpm dev:api & pnpm dev:frontend"
+beforeDevCommand = "pnpm dev:frontend"
 frontendDist = "../studio/dist"
 
 [bundle]
